# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Object Boundary Detection Service",
    "description": "Production-ready Object Boundary Detection Microservice. Extracts exact polygon boundary coordinates and inner holes for objects in photos based on user click points (x, y). Supports local storage, AWS S3, and GCP Cloud Storage.",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "System Health"
        ],
        "summary": "Root",
        "description": "Root endpoint \u2014 returns API metadata and active provider information.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System Health"
        ],
        "summary": "Health Check",
        "description": "Lightweight health probe for Kubernetes, GCP Cloud Run, and AWS ECS.\n\nMemory reading is cached (TTL = HEALTH_CACHE_TTL_SECONDS) to avoid a\npsutil syscall on every frequent load-balancer probe.\n`model_loaded`/`checkpoint_bytes` are file-backed signals only \u2014 they never\ntrigger a model load, so probes stay cheap.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/boundary": {
      "post": {
        "tags": [
          "Boundary Detection"
        ],
        "summary": "Detect object boundary by click point",
        "description": "Detect exact object boundary polygon from an image path and click point (x, y).\n\n- **image_path**: local path, relative filename, `s3://`, `gs://`, or `http://` URL\n- **x** / **y**: click coordinates in pixel space\n- **tolerance**: RDP simplification factor (0.0001\u20130.1, default 0.005)\n- **model_provider**: deprecated, ignored \u2014 MobileSAM-only service.",
        "operationId": "detect_object_boundary_api_v1_boundary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoundaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoundaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boundary/embedding/{photo_id}": {
      "delete": {
        "tags": [
          "Boundary Detection"
        ],
        "summary": "Delete a cached embedding for a photo",
        "operationId": "delete_embedding_api_v1_boundary_embedding__photo_id__delete",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boundary/encode": {
      "post": {
        "tags": [
          "Boundary Detection"
        ],
        "summary": "Pre-calculate embeddings for an image",
        "operationId": "encode_image_endpoint_api_v1_boundary_encode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BoundaryRequest": {
        "properties": {
          "photo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id",
            "description": "Unique ID of the photo (if pre-encoded)"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path",
            "description": "Local file path ('sample.jpg'), relative name, s3:// URI, gs:// URI, or HTTP presigned URL",
            "example": "sample_hat.png"
          },
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X",
            "description": "Click X coordinate on the target object",
            "example": 300
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y",
            "description": "Click Y coordinate on the target object",
            "example": 200
          },
          "tolerance": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.1,
                "minimum": 0.0001
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerance",
            "description": "Polygon simplification tolerance (Ramer-Douglas-Peucker algorithm). Range: 0.001 to 0.05",
            "default": 0.005,
            "example": 0.005
          },
          "model_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider",
            "description": "Deprecated, ignored \u2014 MobileSAM-only service.",
            "example": "mobile_sam"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "Granularity level: 0 (Fine), 1 (Medium), 2 (Coarse), or None for Auto",
            "default": 0,
            "example": 0
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "BoundaryRequest"
      },
      "BoundaryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if object boundary was successfully identified",
            "default": true
          },
          "outer_boundary": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "type": "array",
            "title": "Outer Boundary",
            "description": "List of (x,y) polygon vertices following the exact shape of the object",
            "example": [
              {
                "x": 180,
                "y": 260
              },
              {
                "x": 230,
                "y": 160
              },
              {
                "x": 370,
                "y": 160
              },
              {
                "x": 420,
                "y": 260
              }
            ]
          },
          "holes": {
            "items": {
              "items": {
                "$ref": "#/components/schemas/Point"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Holes",
            "description": "List of inner hole polygon boundaries (e.g. cutouts inside hats, donut holes, mug handles)",
            "default": [],
            "example": [
              [
                {
                  "x": 300,
                  "y": 192
                },
                {
                  "x": 318,
                  "y": 210
                },
                {
                  "x": 282,
                  "y": 210
                }
              ]
            ]
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox",
            "description": "Axis-aligned bounding box around the target object"
          },
          "area_pixels": {
            "type": "integer",
            "title": "Area Pixels",
            "description": "Total area of the object mask in pixels",
            "example": 24944
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "description": "Segmentation confidence score between 0.0 and 1.0",
            "example": 0.88
          },
          "image_dimensions": {
            "$ref": "#/components/schemas/ImageDimensions",
            "description": "Original width and height of the image"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used",
            "description": "Name of the segmentation engine used",
            "example": "MobileSAM"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms",
            "description": "End-to-end processing time in milliseconds",
            "example": 14.5
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message detail if success is False"
          }
        },
        "type": "object",
        "required": [
          "outer_boundary",
          "bounding_box",
          "area_pixels",
          "confidence_score",
          "image_dimensions",
          "model_used",
          "processing_time_ms"
        ],
        "title": "BoundaryResponse"
      },
      "BoundingBox": {
        "properties": {
          "xmin": {
            "type": "integer",
            "title": "Xmin",
            "description": "Minimum X coordinate of object bounding box",
            "example": 160
          },
          "ymin": {
            "type": "integer",
            "title": "Ymin",
            "description": "Minimum Y coordinate of object bounding box",
            "example": 160
          },
          "xmax": {
            "type": "integer",
            "title": "Xmax",
            "description": "Maximum X coordinate of object bounding box",
            "example": 441
          },
          "ymax": {
            "type": "integer",
            "title": "Ymax",
            "description": "Maximum Y coordinate of object bounding box",
            "example": 286
          }
        },
        "type": "object",
        "required": [
          "xmin",
          "ymin",
          "xmax",
          "ymax"
        ],
        "title": "BoundingBox"
      },
      "EncodeRequest": {
        "properties": {
          "photo_id": {
            "type": "string",
            "title": "Photo Id",
            "description": "Unique ID of the photo"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path",
            "description": "Local file path ('sample.jpg'), relative name, s3:// URI, gs:// URI, or HTTP presigned URL",
            "example": "sample_hat.png"
          }
        },
        "type": "object",
        "required": [
          "photo_id",
          "image_path"
        ],
        "title": "EncodeRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageDimensions": {
        "properties": {
          "width": {
            "type": "integer",
            "title": "Width",
            "description": "Image width in pixels",
            "example": 600
          },
          "height": {
            "type": "integer",
            "title": "Height",
            "description": "Image height in pixels",
            "example": 600
          }
        },
        "type": "object",
        "required": [
          "width",
          "height"
        ],
        "title": "ImageDimensions"
      },
      "Point": {
        "properties": {
          "x": {
            "type": "integer",
            "title": "X",
            "description": "X pixel coordinate",
            "example": 300
          },
          "y": {
            "type": "integer",
            "title": "Y",
            "description": "Y pixel coordinate",
            "example": 200
          }
        },
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "Point"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "Boundary Detection",
      "description": "Detect precise object polygon boundary coordinates and inner hole geometries from a single user click point (x, y) on an image."
    },
    {
      "name": "System Health",
      "description": "Health check, runtime metrics, active provider status."
    }
  ]
}